change googleLogin to redux
