+++ src/redux/actions/authActions.js
@@ -1,75 +1,104 @@
 import axios from "axios";
 import { setToken, setUser } from "../reducers/authReducer";
 
-export const getMe = (navigate, navigatePathSuccess, navigatePathError) => 
-    async (dispatch, getState) => {
-        try {
-            let { token } = getState().auth;
+export const getMe =
+  (navigate, navigatePathSuccess, navigatePathError) =>
+  async (dispatch, getState) => {
+    try {
+      let { token } = getState().auth;
 
-            if (!token) return;
+      if (!token) return;
 
-            const response = await axios.get(
-                `${import.meta.env.VITE_API_URL}/api/v1/auth/me`,
-                {
-                    headers: {
-                        Authorization: `Bearer ${token}`,
-                    },
-                }
-            );
-            const { data } = response.data;
-
-            dispatch(setUser(data));
-
-            if (navigatePathSuccess) navigate(navigatePathSuccess);
-
-        } catch (error) {
-            if (axios.isAxiosError(error)) {
-                //if token is not valid
-                if (error.response.status === 401) {
-                    dispatch(logout());
-
-                    if (navigatePathError) navigate(navigatePathError);
-                    return;
-                }
-
-                alert(error?.response?.data?.message);
-                return;
-            }
-
-            alert(error?.message);
+      const response = await axios.get(
+        `${import.meta.env.VITE_API_URL}/api/v1/auth/me`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         }
-    };
+      );
+      const { data } = response.data;
+
+      dispatch(setUser(data));
+
+      if (navigatePathSuccess) navigate(navigatePathSuccess);
+    } catch (error) {
+      if (axios.isAxiosError(error)) {
+        //if token is not valid
+        if (error.response.status === 401) {
+          dispatch(logout());
+
+          if (navigatePathError) navigate(navigatePathError);
+          return;
+        }
+
+        alert(error?.response?.data?.message);
+        return;
+      }
+
+      alert(error?.message);
+    }
+  };
 
 export const login = (email, password, navigate) => async (dispatch) => {
-    try {
-        const response = await axios.post(
-            `${import.meta.env.VITE_API_URL}/api/v1/auth/login`,
-            {
-                email,
-                password,
-            }
-        );
-        const { data } = response.data;
-        const { token } = data;
+  try {
+    const response = await axios.post(
+      `${import.meta.env.VITE_API_URL}/api/v1/auth/login`,
+      {
+        email,
+        password,
+      }
+    );
+    const { data } = response.data;
+    const { token } = data;
 
-        //save our token
-        dispatch(setToken(token));
+    //save our token
+    dispatch(setToken(token));
 
-        //redirect to home
-        navigate("/");
-    } catch (error) {
-        if (axios.isAxiosError(error)) {
-            alert(error?.response?.data?.message);
-            return;
-        }
-        alert(error?.message);
+    //redirect to home
+    navigate("/");
+  } catch (error) {
+    if (axios.isAxiosError(error)) {
+      alert(error?.response?.data?.message);
+      return;
     }
+    alert(error?.message);
+  }
 };
-    
+
+export const registerLoginWithGoogleAction =
+  (accesstoken, navigate) => async (dispatch) => {
+    try {
+      let data = JSON.stringify({
+        access_token: accesstoken,
+      });
+
+      let config = {
+        method: "post",
+        maxBodyLength: Infinity,
+        url: `${import.meta.env.VITE_API_URL}/api/v1/auth/google`,
+        headers: {
+          "Content-Type": "application/json",
+        },
+        data: data,
+      };
+
+      const response = await axios.request(config);
+      const { token } = response.data.data;
+
+      dispatch(setToken(token));
+
+      navigate("/");
+    } catch (error) {
+      if (axios.isAxiosError(error)) {
+        alert(error.response.data.message);
+        return;
+      }
+      alert(error.message);
+    }
+  };
+
 export const logout = () => (dispatch) => {
-    dispatch(setToken(null));
-    dispatch(setUser(null));
+  dispatch(setToken(null));
+  dispatch(setUser(null));
 };
-
-
-
